requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
apscheduler==3.6.3
    # via python-telegram-bot
backports.zoneinfo==0.2.1
    # via tzlocal
cachetools==4.2.2
    # via python-telegram-bot
certifi==2021.10.8
    # via python-telegram-bot
dynaconf==3.1.7
    # via -r requirements.in
pylibftdi==0.19.0
    # via -r requirements.in
python-telegram-bot==13.7
    # via -r requirements.in
pytz==2021.3
    # via
    #   apscheduler
    #   python-telegram-bot
six==1.16.0
    # via apscheduler
tornado==6.1
    # via python-telegram-bot
tzlocal==3.0
    # via apscheduler

# The following packages are considered to be unsafe in a requirements file:
# setuptools
